[levanter] Preserve Pallas convolution gradient dtypes
Keep ShortConv parameters in fp32 across bf16 training and return custom-VJP cotangents in their primal dtypes. This prevents mixed-precision tracing from failing before TPU compilation.

Part of #8377.

+++ experiments/grug/moe_pallas_sconv/model.py
@@ -290,7 +290,7 @@ class ShortConv(eqx.Module):
         attention_mask = None if segment_ids is None else segment_ids >= 0
         output, _ = depthwise_causal_convolution(
             input=x,
-            weight=self.weight.astype(x.dtype),
+            weight=self.weight,
             attention_mask=attention_mask,
         )
